Moved the api-types file and removed unused code
